Moved controllers to components folder
