+++ frontend/client/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App.js';
-import StaffRecordsStage from './actions/StaffRecordsDataGridController';
-import CustomerReservationsStage from './actions/CustomerReservationsController';
+import StaffRecordsStage from './components/staff-records/StaffRecordsDataGridController';
+import CustomerReservationsStage from './components/customerReservations/CustomerReservationsController';
 
 ReactDOM.render(<App />, document.getElementById('root'));
